jelly works and security realm is selectable in jenkins.
The trick was to run the launcher with 'clean hpi:run' to show the new module.

Now we can setup the oauth application details (well we hard code to the test account for now)

And the login is delegated properly to the GitHubSecurityRealm.

Next step is to try out binding to signpost-oauth to see if it will work.

Also changed the licence to the MIT licence since that is what I have been copying from to get to this point.