implement github authorization strategy
This is the top level access strategy.  We allow a comma delimited list of usernames to be given for the administrator users.

We also have a check box that can be used to let authenticated users have a read only view.

This would allow any github user to view the jobs and settings.