Use new command from github-api to get list of organizations of a user.
There are some problems with the github API that are preventing me from getting the list of team members of an organization (for some reason the user requesting this is left out from the returnded lists).

So I have adopted a broader approach where we will just check if the authenticated user is a member of a specific organization and if so grant them access to viewing and running the builds.