Add in test case with api details externalized through Systetm Properties
This is the test code I have been using to figure out how to get the members list of an organization that the current user belongs to.

Because of the nature of the login api key or the oauth key they can't be included in the source code so I have externalized the fields so others can still run the code.