Fix display of authorization strategy properties
Because I was using a List<String> I couldn't get the String fields to fill with existing content.

The fix is to store the comma delmited strings in the GithubAuthorizationStrategy itself and only parse it for the when the ACL is built.

Also bumped up the jenkins version to 1.420 and moved some of the help around (still not working).