Remove build token support in favor of github-plugin webhook.
This commit implements a way to allow anonymous READS for the /github-webhook URL that is supported by the github-plugin.

There is a boolean parameter in the configuration that will enable this url.

This seems to work for a GET request but needs to be tested by doing a push into the monitored github repostory.