Split out ACL from Authorization Strategy.
This splits out the actual ACL from the Authorization Strategy.

At first this was an attempt to isolate a serialization issue but it turned out that this changed the naming that the jelly wiring expected and prevented the existing data to be resolved properly.

This commit uses delegate methods and some string conversion to let the details be shown properly.