Restrict default OAuthScope to "read:org"
The default scope for OAuth should arguably be as restricted as possible. The
older default of "repo,read:org" requested push access to all of a user's
repositories. For many users of the plugin, this was considered a deal-breaker
before configurable OAuth Scopes.
For admins who prefer setting these values through the web UI, it's better to
start with a more restricted scope, as it means that the admin doesn't need to
share push privileges with Jenkins at any point during server configuration.
Plus, this is a better default for novice users who are not familiar with the
GitHub OAuth docs and OAuth scoping.