Allow Jenkins email to be set using private email
Update default OAuth scope to read:org,user:email.  This allows the
Jenkins profile email to be set to the private primary GitHub email.

user:email scope is optional.  If an admin wishes to further restrict
it they can.  It will fall back to attempting to read a public email.

fixes [JENKINS-27764]