Delegate CLI auth to super when no username passed
Attempts to resovle [JENKINS-43822].

See https://issues.jenkins-ci.org/browse/

Since the immediate bug is that CLI behavior is not preserved,
and the existing method override assumes that username is passed
as to the CLI, this patch merely delegates to the superclass when
that required username is not provided.

I don't feel this is an ideal solution to the override-delegate
issue.  The overriding method could attempt to obtain GitHub OAuth
authentication by meeting requirements, then delegate to the
superclass method if OAuth didn't return credentials.