Update log level for private repositories
When a Jenkins user does not have access to certain GitHub repositories (for example private repositories), an exception is thrown.
This reduces the log level to FINEST for such condition, as this is a normal behaviour and not a cause of alarm. 
For a Jenkins instance with many private repositories, logs are filled with those errors.