BranchJobProperty is set only on workflowjobs created by multibranch
So in order to make multibranch build cancellable we also need to take
those jobs into account. Those jobs don't have SCMs setup so we need to
get the scm from the BranchJobProperty like it was done before.

So, now we check if the BranchJobProperty is set and do the adequate
commands to get the scm.

Also, added tests for CANCEL permission which was never tested.