[JENKINS-57154] Fix configureSecurity HTTP 403 err
If an admin visits the `configureSecurity` page in Jenkins, then every
user queried will attempt to be impersonated as part of determining if
they're a user.  However, it's possible for some users to revoke the
OAuth app or no longer have access so impersonation is not possible due
to an invalid token.

The `GithubAuthenticationToken` class did not properly surface an error
when a token authentication was not valid.

See also:

- [JENKINS-57154][JENKINS-57154] Regression in github-oauth-plugin 0.32
  breaks /configureSecurity page

[JENKINS-57154]: https://issues.jenkins-ci.org/browse/JENKINS-57154