Update GithubSecurityRealm.java for consistency in URL building.
There is an open PR to facilitate configuration of redirect_uri but it appears to me that it only hit one of the 2 places that the uri needed to be appended. I am modifying the code for hygiene to make it more obvious where to slip that in (and perhaps also slip in allow_signup).