[JENKINS-34835] Use team slug for authorities if set
PR #116 loads the team based on the slug if it's available but the
authorities list is still using the team name.
This means that you can set the team in the matrix in global security
but you will then be denied access because the authorities list uses the
team name.
You can see this if you go to /whoAmI The team name is shown against the
user.