[JENKINS-34835] Add slug to teams
The GithubAuthenticationToken constructor now uses the github team slug
if it's set. I've updated the tests so the slug is set here as well.

I haven't included a separate test for having a team slug set vs not
having it set but I can add that in if it's needed.