[JENKINS-34835] Authorities from team slug (#124)
* [JENKINS-34835] Use team slug for authorities if set

PR #116 loads the team based on the slug if it's available but the
authorities list is still using the team name.
This means that you can set the team in the matrix in global security
but you will then be denied access because the authorities list uses the
team name.
You can see this if you go to /whoAmI The team name is shown against the
user.

* [JENKINS-34835] Add slug to teams

The GithubAuthenticationToken constructor now uses the github team slug
if it's set. I've updated the tests so the slug is set here as well.