Fix the usage of proxies which require authentication
If the configured HTTP proxy requires basic authentication then the Username and Password
which are configured on the "Advanced" section in "Manage Plugins" section are not used
to route traffic. This leads to HTTP 407 error when trying to connect to the Github or
enterprise Git domain.

This commit ensures that if the username and proxy are configured for the currently configured
HTTP_PROXY that these are also used. The solution has been manually tested, but no unit tests
have been provided.

This commit fixes [JENKINS-67383]