Fix the usage of github authentication plugin behind proxies requiring authentication
If we configure the Jenkins server to be used behind a proxy which requires
authentication, then Jenkins authentication plugin becomes unusable. The
reason for that is the fact that configured username and password are not
passed into the httpclient from Jenkins configuration.

To reproduce this error do following:

- Start Jenkins server behind proxy requiring authentication
- Configure the gihub authentication plugin
- Clear your browser cookies
- Try to login
- Observe the login looping due to missing access token

This commit fixes the issue.