Do not check gh in GithubAuthenticationToken methods
In https://github.com/jenkinsci/github-oauth-plugin/pull/61, use of `gh`
was replaced by `getGitHub()` but the check for `gh != null` was not
removed. `getGitHub()` already ensures that it's present so there's no
need to do the check.

The check was causing calls to `loadUser`, etc to return `null` if a
token was restored and nothing that transitively calls `getGitHub()` has
been called yet.

In particular, when a user logs in, their authorities will be empty
because GithubSecurityRealm tries to `loadUser` in `loadUserByUsername2`
and sets authorities to empty if the user is `null`. Only later, when
something transitively calls `getGitHub()`, will the user have the
authorities set.